push 06/12

+++ src/main/java/com/demo/controllers/user/PlaylistUserController.java
@@ -8,9 +8,12 @@ import java.util.List;
 
 import javax.servlet.ServletContext;
 
+import org.apache.jasper.tagplugins.jstl.core.If;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.util.MimeTypeUtils;
@@ -22,14 +25,18 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.context.ServletContextAware;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.demo.entities.Account;
 import com.demo.entities.Playlist;
 import com.demo.entities.PlaylistCategory;
 import com.demo.entities.Status;
 import com.demo.entities.Track;
 import com.demo.helpers.FileUploadHelper;
+import com.demo.models.AlbumInfo;
 import com.demo.models.TrackInfo;
 import com.demo.services.AccountPlaylistService;
+import com.demo.services.AccountService;
 import com.demo.services.ArtistTrackService;
+import com.demo.services.CookieService;
 import com.demo.services.PlaylistService;
 import com.demo.services.PlaylistTrackService;
 
@@ -50,136 +57,262 @@ public class PlaylistUserController implements ServletContextAware{
 	
 	@Autowired
 	private PlaylistTrackService playlistTrackService;
+	
+	@Autowired
+	private CookieService cookieService;
+	
+	@Autowired
+	private AccountService accountService;
 
 	@RequestMapping( value = {"","index", "manage" })
-	public String index(ModelMap modelMap) {
-		int artistId = 5;
-		modelMap.put("albums", accountPlaylistService.getAlbumsByArtistId(artistId));
+	public String index(ModelMap modelMap, Authentication authentication) {
+		if( authentication == null || authentication instanceof AnonymousAuthenticationToken) {
+			modelMap.put("msg", "You must sign in first") ; 
+			return "redirect:/login/login" ; 
+		}
+		else {			
+			Account account = new Account() ; 
+			String id = cookieService.getValue("acc_id", "") ; 
+			if( id.equalsIgnoreCase("")) {
+				return "redirect:/login/login" ; 
+			}
+			else {
+				account = accountService.findById(Integer.parseInt(id)) ;					
+			}
+			if( account != null ) {					
+				if(!account.isIsArtist()) {
+					modelMap.put("msg", "You not allow to be here!"); 
+					return "redirect:/home" ;
+				}else {
+					modelMap.put("accountSignined", account) ; 
+					int artistId = account.getId();
+					modelMap.put("albums", accountPlaylistService.getAlbumsByArtistId(artistId));
+				}			
+			}
+		} 
 		return "playlist/manage" ; 
 	}
 	
 	@RequestMapping( value = { "add" })
-	public String add(ModelMap modelMap) {
-		int artistId = 5;
-		modelMap.put("tracks", artistTrackService.getTracksOfArtist(artistId));
+	public String add(ModelMap modelMap, Authentication authentication) {
+		if( authentication == null || authentication instanceof AnonymousAuthenticationToken) {
+			modelMap.put("msg", "You must sign in first") ; 
+			return "redirect:/login/login" ; 
+		}
+		else {			
+			Account account = new Account() ; 
+			String id = cookieService.getValue("acc_id", "") ; 
+			if( id.equalsIgnoreCase("")) {
+				return "redirect:/login/login" ; 
+			}
+			else {
+				account = accountService.findById(Integer.parseInt(id)) ;					
+			}
+			if( account != null ) {					
+				if(!account.isIsArtist()) {
+					modelMap.put("msg", "You not allow to be here!"); 
+					return "redirect:/home" ;
+				}else {
+					modelMap.put("accountSignined", account) ; 
+					int artistId = account.getId();
+					modelMap.put("tracks", artistTrackService.getTracksOfArtist(artistId));
+				}			
+			}
+		} 
 		return "playlist/add" ; 
 	}
 	
 	@RequestMapping( value = { "add" }, method = RequestMethod.POST)
-	public String add(@ModelAttribute("playlist") Playlist playlist, @RequestParam(value = "albumDate", required = false) String publishDate, @RequestParam("thumbnailAlbum") MultipartFile thumbnailAlbum, @RequestParam(value = "tracks", required = false) String[] tracks) {
-		int artistId = 5;
-		FileUploadHelper fileHelper = new FileUploadHelper();
-		SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd");
-		String directionThumbnail = "images/playlist/";
-		
-		String thumbnail = fileHelper.uploadFile(thumbnailAlbum, directionThumbnail, servletContext);
-		if(thumbnail != null) {
-			PlaylistCategory playlistCategory = new PlaylistCategory();
-			playlistCategory.setId(3);
-			
-			Status status = new Status();
-			status.setId(1);
-			status.setId(3);
-			
-			Playlist newAlbum = new Playlist();
-			newAlbum.setPlaylistCategory(playlistCategory);
-			newAlbum.setStatus(status);
-			newAlbum.setTitle(playlist.getTitle());
-			newAlbum.setThumbnail(thumbnail);
-			if(publishDate != "") {
-				try {
-					newAlbum.setPublishDate(simpleDateFormat.parse(publishDate));
-				} catch (ParseException e) {
-					e.printStackTrace();
-				}
-			}else {
-				newAlbum.setPublishDate(new Date());
-			}
-			newAlbum.setLastUpdated(new Date());
-			newAlbum.setDescription(playlist.getDescription());
-			newAlbum.setLikes(0);
-			
-			Playlist albumTmp = playlistService.save(newAlbum);
-			if(albumTmp.getId() != 0) {
-				
-				// Set album owner
-				accountPlaylistService.setOwnerAlbum(albumTmp, artistId);
-				
-				// Set tracks for album
-				if(tracks != null) {
-					for(String trackIdString : tracks) {
-						int trackId = Integer.parseInt(trackIdString);
-						playlistTrackService.addTrackToAlbum(albumTmp.getId(), trackId);
-					}		
-				}
-				
-			}			
+	public String add(@ModelAttribute("playlist") Playlist playlist, @RequestParam(value = "albumDate", required = false) String publishDate, @RequestParam("thumbnailAlbum") MultipartFile thumbnailAlbum, @RequestParam(value = "tracks", required = false) String[] tracks, ModelMap modelMap, Authentication authentication) {
+		if( authentication == null || authentication instanceof AnonymousAuthenticationToken) {
+			modelMap.put("msg", "You must sign in first") ; 
+			return "redirect:/login/login" ; 
 		}
+		else {			
+			Account account = new Account() ; 
+			String id = cookieService.getValue("acc_id", "") ; 
+			if( id.equalsIgnoreCase("")) {
+				return "redirect:/login/login" ; 
+			}
+			else {
+				account = accountService.findById(Integer.parseInt(id)) ;					
+			}
+			if( account != null ) {					
+				if(!account.isIsArtist()) {
+					modelMap.put("msg", "You not allow to be here!"); 
+					return "redirect:/home" ;
+				}else {
+					modelMap.put("accountSignined", account) ; 
+					int artistId = account.getId();
+					FileUploadHelper fileHelper = new FileUploadHelper();
+					SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd");
+					String directionThumbnail = "images/playlist/";
+					
+					String thumbnail = fileHelper.uploadFile(thumbnailAlbum, directionThumbnail, servletContext);
+					if(thumbnail != null) {
+						PlaylistCategory playlistCategory = new PlaylistCategory();
+						playlistCategory.setId(3);
+						
+						Status status = new Status();
+						status.setId(1);
+						status.setId(3);
+						
+						Playlist newAlbum = new Playlist();
+						newAlbum.setPlaylistCategory(playlistCategory);
+						newAlbum.setStatus(status);
+						newAlbum.setTitle(playlist.getTitle());
+						newAlbum.setThumbnail(thumbnail);
+						if(publishDate != "") {
+							try {
+								newAlbum.setPublishDate(simpleDateFormat.parse(publishDate));
+							} catch (ParseException e) {
+								e.printStackTrace();
+							}
+						}else {
+							newAlbum.setPublishDate(new Date());
+						}
+						newAlbum.setLastUpdated(new Date());
+						newAlbum.setDescription(playlist.getDescription());
+						newAlbum.setLikes(0);
+						
+						Playlist albumTmp = playlistService.save(newAlbum);
+						if(albumTmp.getId() != null) {
+							
+							// Set album owner
+							accountPlaylistService.setOwnerAlbum(albumTmp, artistId);
+							
+							// Set tracks for album
+							if(tracks != null) {
+								for(String trackIdString : tracks) {
+									int trackId = Integer.parseInt(trackIdString);
+									playlistTrackService.addTrackToAlbum(albumTmp.getId(), trackId);
+								}		
+							}							
+						}			
+					}
+				}			
+			}
+		} 
 		
 		return "redirect:/playlist/manage" ; 
 	}
 	
 	
 	@RequestMapping( value = { "edit/{id}" })
-	public String edit(ModelMap modelMap, @PathVariable("id") int albumId) {
-		int artistId = 5;
-		Playlist album = playlistService.find(albumId);
-		modelMap.put("tracks", artistTrackService.getTracksOfArtist(artistId));
-		modelMap.put("chooseTracks", album.findTracks());
-		modelMap.put("albumId", album.getId());
-		modelMap.put("playlist", album);
-		modelMap.put("thumbnail", album.getThumbnail());
-		modelMap.put("statusId", album.getStatus().getId());
-		modelMap.put("albumDate", album.getPublishDate());
-		modelMap.put("albumId", album.getId());
+	public String edit(ModelMap modelMap, @PathVariable("id") int albumId, Authentication authentication) {
+		if( authentication == null || authentication instanceof AnonymousAuthenticationToken) {
+			modelMap.put("msg", "You must sign in first") ; 
+			return "redirect:/login/login" ; 
+		}
+		else {			
+			Account account = new Account() ; 
+			String id = cookieService.getValue("acc_id", "") ; 
+			if( id.equalsIgnoreCase("")) {
+				return "redirect:/login/login" ; 
+			}
+			else {
+				account = accountService.findById(Integer.parseInt(id)) ;					
+			}
+			if( account != null ) {					
+				if(!account.isIsArtist()) {
+					modelMap.put("msg", "You not allow to be here!"); 
+					return "redirect:/home" ;
+				}else {
+					modelMap.put("accountSignined", account) ; 
+					int artistId = account.getId();
+					if(accountPlaylistService.checkAlbumOwner(artistId, albumId)) {
+						Playlist album = playlistService.find(albumId);
+						modelMap.put("tracks", artistTrackService.getTracksOfArtist(artistId));
+						modelMap.put("chooseTracks", album.findTracks());
+						modelMap.put("albumId", album.getId());
+						modelMap.put("playlist", album);
+						modelMap.put("thumbnail", album.getThumbnail());
+						modelMap.put("statusId", album.getStatus().getId());
+						modelMap.put("albumDate", album.getPublishDate());
+						modelMap.put("albumId", album.getId());						
+					}else {
+						modelMap.put("msg", "Sorry, but you didn't own this album!"); 
+						return "redirect:/playlist/manage" ;
+					}
+				}			
+			}
+		} 
 		return "playlist/edit" ; 
 	}
 	
 	
 	@RequestMapping( value = { "edit" }, method = RequestMethod.POST)
-	public String edit(@ModelAttribute("playlist") Playlist playlist, @RequestParam("thumbnailAlbum") MultipartFile thumbnailAlbum, @RequestParam(value = "tracks", required = false) String[] tracks, @RequestParam(value = "isHidden", required = false) String isHidden) {
-		int artistId = 5;
-		Playlist album = playlistService.find(playlist.getId());
-		FileUploadHelper fileHelper = new FileUploadHelper();
-		String directionThumbnail = "images/playlist/";
-		
-		// Set new thumbnail
-		if (!thumbnailAlbum.isEmpty() && thumbnailAlbum.getSize() > 0) {
-			String thumbnail = fileHelper.uploadFile(thumbnailAlbum, directionThumbnail, servletContext);
-			if(thumbnail != null) {
-				boolean isDelete = fileHelper.deleteFile(album.getThumbnail(), directionThumbnail, servletContext);
-				if(isDelete) {
-					album.setThumbnail(thumbnail);
-				}				
+	public String edit(@ModelAttribute("playlist") Playlist playlist, @RequestParam("thumbnailAlbum") MultipartFile thumbnailAlbum, @RequestParam(value = "tracks", required = false) String[] tracks, @RequestParam(value = "isHidden", required = false) String isHidden, ModelMap modelMap, Authentication authentication) {
+		if( authentication == null || authentication instanceof AnonymousAuthenticationToken) {
+			modelMap.put("msg", "You must sign in first") ; 
+			return "redirect:/login/login" ; 
+		}
+		else {			
+			Account account = new Account() ; 
+			String id = cookieService.getValue("acc_id", "") ; 
+			if( id.equalsIgnoreCase("")) {
+				return "redirect:/login/login" ; 
 			}
-		}
+			else {
+				account = accountService.findById(Integer.parseInt(id)) ;					
+			}
+			if( account != null ) {					
+				if(!account.isIsArtist()) {
+					modelMap.put("msg", "You not allow to be here!"); 
+					return "redirect:/home" ;
+				}else {
+					modelMap.put("accountSignined", account) ; 
+					int artistId = account.getId();
+					if(accountPlaylistService.checkAlbumOwner(artistId, playlist.getId())) {
+						Playlist album = playlistService.find(playlist.getId());
+						FileUploadHelper fileHelper = new FileUploadHelper();
+						String directionThumbnail = "images/playlist/";
+						
+						// Set new thumbnail
+						if (!thumbnailAlbum.isEmpty() && thumbnailAlbum.getSize() > 0) {
+							String thumbnail = fileHelper.uploadFile(thumbnailAlbum, directionThumbnail, servletContext);
+							if(thumbnail != null) {
+								boolean isDelete = fileHelper.deleteFile(album.getThumbnail(), directionThumbnail, servletContext);
+								if(isDelete) {
+									album.setThumbnail(thumbnail);
+								}				
+							}
+						}
+						
+						// Check/Set publish/hidden
+						Status status = new Status();			
+						if(isHidden != null) {
+							status.setId(3);
+						}else {
+							status.setId(1);
+						}
+						album.setStatus(status);	
+						
+						album.setTitle(playlist.getTitle());
+						album.setDescription(playlist.getDescription());
+						album.setLastUpdated(new Date());
+						
+						// Update album
+						playlistService.save(album);
+						
+						// Delete old tracks from album
+						playlistTrackService.removeAllTrackFromAlbum(album);
+						
+						// Add tracks to album
+						if(tracks != null) {
+							for(String trackIdString : tracks) {
+								int trackId = Integer.parseInt(trackIdString);
+								playlistTrackService.addTrackToAlbum(album.getId(), trackId);
+							}		
+						}		
+					}else {
+						modelMap.put("msg", "Sorry, but you didn't own this album!"); 
+						return "redirect:/playlist/manage" ;
+					}
+				}			
+			}
+		} 
 		
-		// Check/Set publish/hidden
-		Status status = new Status();			
-		if(isHidden != null) {
-			status.setId(3);
-		}else {
-			status.setId(1);
-		}
-		album.setStatus(status);	
-		
-		album.setTitle(playlist.getTitle());
-		album.setDescription(playlist.getDescription());
-		album.setLastUpdated(new Date());
-		
-		// Update album
-//		playlistService.save(album);
-		
-		// Delete old tracks from album
-		playlistTrackService.removeAllTrackFromAlbum(album);
-		
-		// Add tracks to album
-		if(tracks != null) {
-			for(String trackIdString : tracks) {
-				int trackId = Integer.parseInt(trackIdString);
-				playlistTrackService.addTrackToAlbum(album.getId(), trackId);
-			}		
-		}		
 		
 		return "redirect:/playlist/manage" ; 
 	}
@@ -205,22 +338,69 @@ public class PlaylistUserController implements ServletContextAware{
 		}
 	}
 	
-	@RequestMapping(value = { "delete" }, method = RequestMethod.GET, produces = MimeTypeUtils.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Boolean> delete(@RequestParam("id") int albumId) {
-		boolean result = false;
-		Playlist album = playlistService.find(albumId);
-		
-		accountPlaylistService.removeAccountHasAlbum(album);
-		
-		playlistTrackService.removeAllTrackFromAlbum(album);
-		
-		playlistService.delete(albumId);
-		
-		result = true;
+	@RequestMapping(value = { "delete" }, method = RequestMethod.GET, produces = MimeTypeUtils.TEXT_PLAIN_VALUE)
+	public ResponseEntity<String> delete(@RequestParam("id") int albumId, ModelMap modelMap, Authentication authentication) {
+		String result = "";
+		if( authentication == null || authentication instanceof AnonymousAuthenticationToken) {
+			result = "MUST_SIGN_IN";
+		}
+		else {			
+			Account account = new Account() ; 
+			String id = cookieService.getValue("acc_id", "") ; 
+			if( id.equalsIgnoreCase("")) {
+				result = "MUST_SIGN_IN";
+			}
+			else {
+				account = accountService.findById(Integer.parseInt(id)) ;					
+			}
+			if( account != null ) {					
+				if(!account.isIsArtist()) {
+					result = "NOT_ALLOW";
+				}else {
+					modelMap.put("accountSignined", account) ; 
+					int artistId = account.getId();
+					if(accountPlaylistService.checkAlbumOwner(artistId, albumId)) {
+						Playlist album = playlistService.find(albumId);
+						
+						accountPlaylistService.removeAccountHasAlbum(album);
+						
+						playlistTrackService.removeAllTrackFromAlbum(album);
+						
+						playlistService.delete(albumId);
+						
+						result = "OK";					
+					}
+				}			
+			}
+		} 
+
 		try {
-			return new ResponseEntity<Boolean>(result, HttpStatus.OK);
+			return new ResponseEntity<String>(result, HttpStatus.OK);
 		} catch (Exception e) {
-			return new ResponseEntity<Boolean>(HttpStatus.BAD_REQUEST);
+			return new ResponseEntity<String>("ERROR", HttpStatus.BAD_REQUEST);
+		}
+	}
+	
+	// [Manage] Search
+	@RequestMapping(value = { "searchAlbumInManage" }, method = RequestMethod.GET, produces = MimeTypeUtils.APPLICATION_JSON_VALUE)
+	public ResponseEntity<List<AlbumInfo>> searchTrackInManage(@RequestParam("option") String option, @RequestParam("keyword") String keyword) {
+		List<AlbumInfo> albumInfos = new ArrayList<AlbumInfo>();
+		Account account = new Account() ; 
+		String id = cookieService.getValue("acc_id", "") ;
+		String optionValue = "";
+		if(option.equalsIgnoreCase("all")) {
+			optionValue = "";
+		}else {
+			optionValue = option;
+		}
+		if(!id.equalsIgnoreCase("")) {
+			account = accountService.findById(Integer.parseInt(id));
+			albumInfos = playlistService.searchAlbumInManage(optionValue, keyword.toLowerCase(), account.getId());
+		}
+		try {
+			return new ResponseEntity<List<AlbumInfo>>(albumInfos, HttpStatus.OK);
+		} catch (Exception e) {
+			return new ResponseEntity<List<AlbumInfo>>(HttpStatus.BAD_REQUEST);
 		}
 	}
